Task/15 create chart of breakdown of coding over the last 7 days by each day

+++ iOS-client-wakatime/Models/EntrySummary.swift
@@ -12,14 +12,14 @@ import ObjectMapper
 class EntrySummary: Mappable {
     
     var entryName: String?
-    var workingTimeInPercent: Float?
+    var workingTimeInPercent: Double?
     var workingTimeAsText: String?
     var totalWorkingTimeAsSecond: Int?
     var workingTimeHoursPortion: Int?
     var workingTimeMinutesPortion: Int?
     
     init(entryName: String,
-         workingTimeInPercent: Float,
+         workingTimeInPercent: Double,
          workingTimeAsText: String,
          totalWorkingTimeAsSecond: Int,
          workingTimeHoursPortion: Int,
